+++ docket/bot.py
@@ -14,7 +14,8 @@ class Docket(crescent.Bot):
 
         self.plugins.load("docket.plugins.info")
         self.plugins.load("docket.plugins.events")
-        self.plugins.load("docket.plugins.config")
+        self.plugins.load("docket.plugins.config_scripts")
+        self.plugins.load("docket.plugins.config_event_triggers")
         self.plugins.load("docket.plugins.error")
 
         self.sync_async = SyncAsync()
